completed delete action feature

+++ src/components/Product.js
@@ -1,9 +1,10 @@
 import React from 'react'
 import {Link} from 'react-router-dom'
-import { deleteProductAction } from '../actions/productActions'
-
+import Swal from 'sweetalert2';
 // Redux
 import { useDispatch } from 'react-redux'
+import { deleteProductAction } from '../actions/productActions'
+
 const Product = ({product}) => {
     const {name,price,id} = product
     
@@ -12,9 +13,22 @@ const Product = ({product}) => {
     // Delete confirmation
     const productDeleteConfirmation = id => {
         // ask user
+        Swal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Delete',
+            cancelButtonText: 'Cancel'
+          }).then((result) => {
+            if (result.isConfirmed) {
+                // send to action
+                dispatch(deleteProductAction(id));
+            }
+          })
 
-        // send to action
-        dispatch(deleteProductAction(id));
     }
     
     return (
